+++ src/backend/core/services/ai_services.py
@@ -31,6 +31,11 @@ AI_ACTIONS = {
         "Do not provide any other information. "
         "Preserve the language."
     ),
+    "beautify": (
+        "Add formatting to the text to make it more readable. "
+        "Do not provide any other information."
+        "Preserve the language."
+    ),
 }
 
 AI_TRANSLATE = (