+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/BlockNoteToolBar/AIButton.tsx
@@ -165,6 +165,17 @@ export function AIGroupButton() {
             >
               {t('Correct')}
             </AIMenuItemTransform>
+            <AIMenuItemTransform
+              action="beautify"
+              docId={currentDoc.id}
+              icon={
+                <Text $isMaterialIcon $size="s">
+                  draw
+                </Text>
+              }
+            >
+              {t('Beautify')}
+            </AIMenuItemTransform>
           </>
         )}
         {canAITranslate && (